pyproject.toml
[project]
name = "perplexity-mcp"
version = "0.4.0"
description = "Free Perplexity AI for your coding assistant. No API keys, no subscription."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["perplexity", "ai", "mcp", "search"]

dependencies = [
    # MCP (stdio server)
    "mcp>=1.0.0",
    # Backend
    "curl_cffi>=0.5.7",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    # HTTP server
    "starlette>=0.27.0",
    "uvicorn>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
perplexity-mcp = "src.server:main"
perplexity-server = "perplexity.server.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "perplexity"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
